Added an additional lower threshold temperature for use during away mode
The devices can't actually be set to this temperature, as it is outside the valid temperature setpoint range. Instead, this value is used as a trigger for the device to enter heating mode with the valid minimum temperature as the setpoint. Then, when the room temperature rises above this minimum threshold, the device enters fan only mode. In that way, it will mimic the behavior of a lower minimum setpoint than is allowed by the device.